some more cleaning up - the exposed errors will need to be fixed

+++ src/common/IGrid.java
@@ -20,7 +20,7 @@ public interface IGrid {
 	
 	public float getPlanetY();	
 	
-	public int getCurrentTime();
+	public long getCurrentTime();
 	
 	public int getGridWidth();
